refactor: integrate serializer modules
