--- app/serializers/CL_list_serializer.py
@@ -1,7 +1,0 @@
-from ..models import CLData
-from rest_framework import serializers
-
-class CLListSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = CLData
-        fields = ["id", "title", "content"]
