+++ app/serializers/__init__.py
@@ -1,29 +1,5 @@
-from .user_serializer import UserSerializer
-
-from .CL_serializer import CLSerializer
-from .CL_list_serializer import CLListSerializer
-
-from .position_serializer import PositionSerializer
-from .position_list_serializer import PositionListSerializer
-from .position_detail_serializer import PositionDetailSerializer
-
-from .resume_serializer import ResumeSerializer
-
-from .claim_serializer import ClaimSerializer
-from .claim_list_serializer import EmployerClaimListSerializer, EmployeeClaimListSerializer
-
-# Ignore unused
-_ = UserSerializer
-
-_ = CLSerializer
-_ = CLListSerializer
-
-_ = PositionSerializer
-_ = PositionListSerializer
-_ = PositionDetailSerializer
-
-_ = ResumeSerializer
-
-_ = ClaimSerializer
-_ = EmployerClaimListSerializer
-_ = EmployeeClaimListSerializer
+from .user_serializer       import *
+from .CL_serializer         import *
+from .position_serializer   import *
+from .resume_serializer     import *
+from .claim_serializer      import *
